Draw code documentation from decomp and pointer truncation fixups

+++ src/BGS/ch/croctus.c
@@ -91,7 +91,7 @@ void func_80387E00(s32 arg0){
     Actor * this = marker_getActor(marker);
 
     func_803262E4(this);
-    __spawnQueue_add_2((GenFunction_2)func_80387D18, reinterpret_cast(s32, marker), 0x1E);
+    __spawnQueue_add_2((GenFunction_2)func_80387D18, reinterpret_cast(uintptr_t, marker), 0x1E);
 }
 
 void func_80387E40(ActorMarker * arg0){
@@ -112,12 +112,12 @@ void chCroctus_activateNextCroctus(ActorMarker *caller, enum asset_e text_id, s3
         subaddie_set_state_with_direction(this, CHCROCTUS_STATE_5_FED, 0.79f, 1);
         func_80326310(this); //did not disappear when moved, after cutscene still there with collision but broken
         bgs_D_803907B8[this->actorTypeSpecificField]->propPtr->isNotFeatherEggOrNote = TRUE;
-        timedFunc_set_1(1.1f, (GenFunction_1)func_80387E00, reinterpret_cast(s32, bgs_D_803907B8[this->actorTypeSpecificField]));
+        timedFunc_set_1(1.1f, (GenFunction_1)func_80387E00, reinterpret_cast(uintptr_t, bgs_D_803907B8[this->actorTypeSpecificField]));
         if(!sCroctusRemote){
             timed_setStaticCameraToNode(0.8f, 9);
             func_80324DBC(3.4f, VER_SELECT(ASSET_C87_DIALOG_CROCTUS_SECOND_SPAWN, 0x9CA, 0, 0), 0xE, NULL, NULL, chCroctus_activateNextCroctus, NULL);
         }
-        __spawnQueue_add_2((GenFunction_2) func_80387D18, reinterpret_cast(s32, this->marker), 0x46);
+        __spawnQueue_add_2((GenFunction_2) func_80387D18, reinterpret_cast(uintptr_t, this->marker), 0x46);
     }
     else{
         timed_exitStaticCamera(0.0f);
@@ -195,18 +195,18 @@ void chCroctus_update(Actor *this){
                 func_80324D2C(4.5f, COMUSIC_43_ENTER_LEVEL_GLITTER);
                 subaddie_set_state_with_direction(this, CHCROCTUS_STATE_5_FED, 0.79f, 1);
                 if (this->actorTypeSpecificField == 5) {
-                    timedFunc_set_1(0.9f, (GenFunction_1) func_80387E40, reinterpret_cast(s32, this->marker));
+                    timedFunc_set_1(0.9f, (GenFunction_1) func_80387E40, reinterpret_cast(uintptr_t, this->marker));
                 } else {
                     func_80326310(this);
                 }
                 if (this->actorTypeSpecificField < 5) {
                     bgs_D_803907B8[this->actorTypeSpecificField]->propPtr->isNotFeatherEggOrNote = TRUE;
-                    timedFunc_set_1(1.1f, (GenFunction_1)func_80387E00, reinterpret_cast(s32, bgs_D_803907B8[this->actorTypeSpecificField]));
+                    timedFunc_set_1(1.1f, (GenFunction_1)func_80387E00, reinterpret_cast(uintptr_t, bgs_D_803907B8[this->actorTypeSpecificField]));
                     if (!sCroctusRemote) gcStaticCamera_activate(D_803907B0[this->actorTypeSpecificField-1]);
                 } else {
-                    timedFunc_set_1(0.8f, (GenFunction_1)chCroctus_jiggySpawn, reinterpret_cast(s32, this->marker));
+                    timedFunc_set_1(0.8f, (GenFunction_1)chCroctus_jiggySpawn, reinterpret_cast(uintptr_t, this->marker));
                 }
-                __spawnQueue_add_2((GenFunction_2)func_80387D18, reinterpret_cast(s32, this->marker), 0x46);
+                __spawnQueue_add_2((GenFunction_2)func_80387D18, reinterpret_cast(uintptr_t, this->marker), 0x46);
             }
         }
     }//L80388348
